requirements.txt
opencv-python
pytesseract
Pillow
transformers
torch
openai-whisper
